+++ frontend/src/components/LotApplicantAndInfoForm/index.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useQuery, UseQueryResult, useMutation } from '@tanstack/react-query';
 
@@ -6,15 +6,10 @@ import {
   Row,
   Column,
   TextInput,
-  RadioButtonGroup,
-  RadioButton,
-  Checkbox,
-  CheckboxGroup,
   Button,
   ComboBox,
   TextInputSkeleton,
   InlineLoading,
-  RadioButtonSkeleton,
   ActionableNotification
 } from '@carbon/react';
 import { DocumentAdd } from '@carbon/icons-react';
@@ -22,38 +17,38 @@ import validator from 'validator';
 import { toast } from 'react-toastify';
 import { AxiosError } from 'axios';
 
-import Subtitle from '../../../../components/Subtitle';
-import InputErrorText from '../../../../components/InputErrorText';
+import Subtitle from '../Subtitle';
+import InputErrorText from '../InputErrorText';
 
-import { ErrToastOption } from '../../../../config/ToastifyConfig';
-import { FilterObj, filterInput } from '../../../../utils/filterUtils';
-import focusById from '../../../../utils/FocusUtils';
-import { THREE_HALF_HOURS, THREE_HOURS } from '../../../../config/TimeUnits';
+import { ErrToastOption } from '../../config/ToastifyConfig';
+import { FilterObj, filterInput } from '../../utils/filterUtils';
+import focusById from '../../utils/FocusUtils';
 
-import { SeedlotRegFormType, SeedlotRegPayloadType } from '../../../../types/SeedlotRegistrationTypes';
-import SeedlotSourceType from '../../../../types/SeedlotSourceType';
-import ComboBoxEvent from '../../../../types/ComboBoxEvent';
+import { SeedlotRegFormType, SeedlotRegPayloadType } from '../../types/SeedlotRegistrationTypes';
+import SeedlotSourceType from '../../types/SeedlotSourceType';
+import ComboBoxEvent from '../../types/ComboBoxEvent';
 
-import getVegCodes from '../../../../api-service/vegetationCodeAPI';
-import getApplicantAgenciesOptions from '../../../../api-service/applicantAgenciesAPI';
-import { getForestClientLocation } from '../../../../api-service/forestClientsAPI';
-import getSeedlotSources from '../../../../api-service/SeedlotSourcesAPI';
-import { postSeedlot } from '../../../../api-service/seedlotAPI';
-import { LOCATION_CODE_LIMIT } from '../../../../shared-constants/shared-constants';
-import ErrorToast from '../../../../components/Toast/ErrorToast';
+import getApplicantAgenciesOptions from '../../api-service/applicantAgenciesAPI';
+import { getForestClientLocation } from '../../api-service/forestClientsAPI';
+import { postSeedlot } from '../../api-service/seedlotAPI';
+import { LOCATION_CODE_LIMIT } from '../../shared-constants/shared-constants';
+import ErrorToast from '../Toast/ErrorToast';
 
-import ComboBoxPropsType from './definitions';
+import { ComboBoxPropsType, FormProps } from './definitions';
 import {
   applicantAgencyFieldConfig,
-  speciesFieldConfig,
   pageTexts,
   InitialSeedlotFormData
 } from './constants';
 import { convertToPayload } from './utils';
 
 import './styles.scss';
+import SeedlotInformation from './SeedlotInformation';
 
-const CreationForm = () => {
+/**
+ * This component displays a form for seedlot or veglot creation/edition.
+ */
+const LotApplicantAndInfoForm = ({ isSeedlot, isEdit }: FormProps) => {
   const navigate = useNavigate();
 
   const [formData, setFormData] = useState<SeedlotRegFormType>(InitialSeedlotFormData);
@@ -94,14 +89,8 @@ const CreationForm = () => {
 
   const applicantAgencyQuery = useQuery({
     queryKey: ['applicant-agencies'],
+    enabled: !isEdit,
     queryFn: () => getApplicantAgenciesOptions()
-  });
-
-  const vegCodeQuery = useQuery({
-    queryKey: ['vegetation-codes'],
-    queryFn: () => getVegCodes(true),
-    staleTime: THREE_HOURS, // will not refetch for 3 hours
-    cacheTime: THREE_HALF_HOURS // data is cached 3.5 hours then deleted
   });
 
   const setDefaultSource = (sources: SeedlotSourceType[]) => {
@@ -117,24 +106,6 @@ const CreationForm = () => {
       }
     });
   };
-
-  const seedlotSourcesQuery = useQuery({
-    queryKey: ['seedlot-sources'],
-    queryFn: () => getSeedlotSources(),
-    onSuccess: (sources) => setDefaultSource(sources),
-    staleTime: THREE_HOURS,
-    cacheTime: THREE_HALF_HOURS
-  });
-
-  /**
-   *  Default value is only set once upon query success, when cache data is used
-   *  we will need to set the default again here.
-   */
-  useEffect(() => {
-    if (seedlotSourcesQuery.isSuccess && !formData.sourceCode.value) {
-      setDefaultSource(seedlotSourcesQuery.data);
-    }
-  }, [seedlotSourcesQuery.isFetched]);
 
   const handleLocationCodeBlur = (clientNumber: string, locationCode: string) => {
     const isInRange = validator.isInt(locationCode, { min: 0, max: 99 });
@@ -204,16 +175,6 @@ const CreationForm = () => {
     }
   };
 
-  const handleSource = (value: string) => {
-    setFormData((prevData) => ({
-      ...prevData,
-      sourceCode: {
-        ...prevData.sourceCode,
-        value
-      }
-    }));
-  };
-
   const handleEmail = (value: string) => {
     const isEmailInvalid = !validator.isEmail(value);
     setFormData((prevData) => ({
@@ -222,16 +183,6 @@ const CreationForm = () => {
         ...prevData.email,
         value,
         isInvalid: isEmailInvalid
-      }
-    }));
-  };
-
-  const handleCheckBox = (inputName: keyof SeedlotRegFormType, checked: boolean) => {
-    setFormData((prevData) => ({
-      ...prevData,
-      [inputName]: {
-        ...prevData[inputName],
-        value: checked
       }
     }));
   };
@@ -252,21 +203,6 @@ const CreationForm = () => {
       search: `?seedlotNumber=${res.data.seedlotNumber}&seedlotClass=A`
     })
   });
-
-  const renderSources = () => {
-    if (seedlotSourcesQuery.isSuccess) {
-      return seedlotSourcesQuery.data.map((source: SeedlotSourceType) => (
-        <RadioButton
-          key={source.code}
-          checked={formData.sourceCode.value === source.code}
-          id={`seedlot-source-radio-btn-${source.code.toLocaleLowerCase()}`}
-          labelText={source.description}
-          value={source.code}
-        />
-      ));
-    }
-    return <InputErrorText description="Could not retrieve seedlot sources." />;
-  };
 
   const displayCombobox = (
     query: UseQueryResult,
@@ -347,7 +283,7 @@ const CreationForm = () => {
   };
 
   return (
-    <div className="applicant-information-form">
+    <Column className="applicant-information-form">
       <form onSubmit={validateAndSubmit}>
         {
           seedlotMutation.isError
@@ -429,65 +365,13 @@ const CreationForm = () => {
             />
           </Column>
         </Row>
-        <Row className="seedlot-information-title">
-          <Column lg={8}>
-            <h2>Seedlot information</h2>
-            <Subtitle text="Enter the initial information about this seedlot" />
-          </Column>
-        </Row>
-        <Row className="seedlot-species-row">
-          {
-            displayCombobox(vegCodeQuery, speciesFieldConfig)
-          }
-        </Row>
-        <Row className="class-source-radio">
-          <Column sm={4} md={8} lg={16}>
-            <RadioButtonGroup
-              legendText="Class A source"
-              name="class-source-radiogroup"
-              orientation="vertical"
-              onChange={(e: string) => handleSource(e)}
-            >
-              {
-                seedlotSourcesQuery.isFetching
-                  ? (
-                    <RadioButtonSkeleton />
-                  )
-                  : renderSources()
-              }
-            </RadioButtonGroup>
-          </Column>
-        </Row>
-        <Row className="registered-checkbox">
-          <Column sm={4} md={8} lg={16}>
-            <CheckboxGroup legendText="To be registered?">
-              <Checkbox
-                id="registered-tree-seed-center"
-                name="registered"
-                labelText="Yes, to be registered with the Tree Seed Centre"
-                checked={formData.willBeRegistered.value}
-                onChange={
-                  (e: React.ChangeEvent<HTMLInputElement>) => handleCheckBox('willBeRegistered', e.target.checked)
-                }
-              />
-            </CheckboxGroup>
-          </Column>
-        </Row>
-        <Row className="collected-checkbox">
-          <Column sm={4} md={8} lg={16}>
-            <CheckboxGroup legendText="Collected from B.C. source?">
-              <Checkbox
-                id="collected-bc"
-                name="collectedBC"
-                labelText="Yes, collected from a location within B.C."
-                checked={formData.isBcSource.value}
-                onChange={
-                  (e: React.ChangeEvent<HTMLInputElement>) => handleCheckBox('isBcSource', e.target.checked)
-                }
-              />
-            </CheckboxGroup>
-          </Column>
-        </Row>
+        {
+          isSeedlot
+            ? (
+              <SeedlotInformation  />
+            )
+            : null
+        }
         <Row className="save-button">
           <Column lg={8}>
             <Button renderIcon={DocumentAdd} type="submit">
@@ -496,8 +380,8 @@ const CreationForm = () => {
           </Column>
         </Row>
       </form>
-    </div>
+    </Column>
   );
 };
 
-export default CreationForm;
+export default LotApplicantAndInfoForm;
