+++ frontend/src/components/LotApplicantAndInfoForm/styles.scss
@@ -2,8 +2,6 @@
 @use '@bcgov-nr/nr-theme/design-tokens/variables.scss' as vars;
 
 .applicant-information-form {
-  margin-top: 2.75rem;
-
   .error-row {
     margin-bottom: 2rem;
