--- frontend/src/views/Seedlot/CreateAClass/CreationForm/definitions.ts
@@ -1,8 +1,0 @@
-type ComboBoxPropsType = {
-  placeholder: string;
-  titleText: string;
-  invalidText: string;
-  helperText: string;
-}
-
-export default ComboBoxPropsType;
